+++ Source/CesiumEditor/Private/CesiumIonSession.cpp
@@ -119,6 +119,7 @@ void CesiumIonSession::disconnect() {
   this->ProfileUpdated.Broadcast();
   this->AssetsUpdated.Broadcast();
   this->TokensUpdated.Broadcast();
+  this->AssetAccessTokenUpdated.Broadcast();
 }
 
 void CesiumIonSession::refreshProfile() {
@@ -237,10 +238,12 @@ void CesiumIonSession::refreshAssetAccessToken() {
       .thenInMainThread([this](CesiumIonClient::Token&& token) {
         this->_assetAccessToken = std::move(token);
         this->_isLoadingAssetAccessToken = false;
+        this->AssetAccessTokenUpdated.Broadcast();
       })
       .catchInMainThread([this](std::exception&& e) {
         this->_assetAccessToken = std::nullopt;
         this->_isLoadingAssetAccessToken = false;
+        this->AssetAccessTokenUpdated.Broadcast();
       });
 }